functions/paragraph-ocr/tesseract/requirements.txt
opencv-python
requests
numpy
pillow
pytesseract